Harmonise gaction object conversion functions
PyYAML is now used for coercing from string any object of type dict or bool, and for converting objects of these types to string. Dictionaries are passed to and from strings in flow-style format, and are passed to and from files in block-style format.

List objects are coerced from a string using PyYAML with flow-style input. Files are converted to a list in which each line contains a list element; each such list element is further parsed using PyYAML.

Each element of a list is serialised to string with PyYAML using flow-style output. If the list is being output to a file, each line of the output file contains one top-level element.

Objects of NoneType are coerced to the string value 'null', and vice versa.