Improve handling of gactfunc return value
If a gactfunc contains an explicit return value, this must be documented in the function docstring, including the return type. The function definition will also be checked during setup, to check for undefined or undocumented return values.

Argument processing now extracts a return value output file, which is set to standard output by default. If a gactfunc returns a value, that value is written to the return value output stream. It can be set to a specific output file using the output flag '-o' on the command line.