Improve gactfunc type handling
Types are now referred to by name, to allow for more flexible type-handling. In particular, the 'string' type corresponds to 'basestring', and parameters of this type should accept either 'str' or 'unicode'.

Types are divided into three subgroups: ductile, delimitable, and compound. Ductile objects can be serialised into a simple single-line string, delimitable objects can be delimited by a newline or other delimiter, and compound types are composed of other data types.

Elements of lists output to a file are stripped of any trailing newline before being output, so as to allow delimitable objects to be output to file as a list element.